refactor(mac): split TerminalHostState concerns
Summary:
- Move TerminalHostState creation, selection, close, catalog, zmx,
  and mapping methods into focused extension files.
- Leave stored state, initialization, dispatch, and unrelated helpers in
  the core TerminalHostState file.

Rationale:
- Keep the Terminal host behavior unchanged while making the large type
  easier to navigate by concern.
- Preserve existing extension-file conventions without starting a deeper
  type decomposition.

Tests:
- make mac-build after each extracted group
- make mac-test after close extraction and final state
- make mac-check
