+++ apps/mac/supaterm/Features/Terminal/TerminalHostState+SpaceCatalog.swift
@@ -1,0 +1,131 @@
+import AppKit
+import Foundation
+import GhosttyKit
+import Observation
+import Sharing
+import SupatermCLIShared
+import SupatermSupport
+import SupatermTerminalCore
+import SwiftUI
+
+extension TerminalHostState {
+  @discardableResult
+  func createSpace(named name: String) throws -> TerminalSpaceID {
+    guard let normalizedName = Self.trimmedNonEmpty(name) else {
+      throw TerminalControlError.invalidSpaceName
+    }
+    guard spaceManager.isNameAvailable(normalizedName) else {
+      throw TerminalControlError.spaceNameUnavailable
+    }
+
+    let space = PersistedTerminalSpace(name: normalizedName)
+    var updatedSpaceCatalog = spaceCatalog
+    updatedSpaceCatalog.defaultSelectedSpaceID = space.id
+    updatedSpaceCatalog = TerminalSpaceCatalog(
+      defaultSelectedSpaceID: updatedSpaceCatalog.defaultSelectedSpaceID,
+      spaces: updatedSpaceCatalog.spaces + [space]
+    )
+    _ = writeSpaceCatalog(updatedSpaceCatalog)
+    guard applySelectedSpace(space.id) else {
+      throw TerminalControlError.spaceNotFound(windowIndex: 1, spaceIndex: spaces.count + 1)
+    }
+    finalizeSpaceSelectionChange()
+    sessionDidChange()
+    return space.id
+  }
+
+  func renameSpace(_ spaceID: TerminalSpaceID, to name: String) {
+    let normalizedName = name.trimmingCharacters(in: .whitespacesAndNewlines)
+    guard !normalizedName.isEmpty else { return }
+    guard spaceManager.isNameAvailable(normalizedName, excluding: spaceID) else { return }
+    guard let index = spaceCatalog.spaces.firstIndex(where: { $0.id == spaceID }) else { return }
+
+    var updatedSpaceCatalog = spaceCatalog
+    updatedSpaceCatalog.spaces[index].name = normalizedName
+    _ = writeSpaceCatalog(updatedSpaceCatalog)
+  }
+
+  func deleteSpace(_ spaceID: TerminalSpaceID) {
+    let remainingSpaces = spaceCatalog.spaces.filter { $0.id != spaceID }
+    guard !remainingSpaces.isEmpty else { return }
+    guard remainingSpaces.count != spaceCatalog.spaces.count else { return }
+    if previousSelectedSpaceID == spaceID {
+      previousSelectedSpaceID = nil
+    }
+
+    let nextSelectedSpaceID = nextSelectedSpaceID(afterDeleting: spaceID)
+    let updatedSpaceCatalog = TerminalSpaceCatalog(
+      defaultSelectedSpaceID: nextSelectedSpaceID,
+      spaces: remainingSpaces
+    )
+    _ = writeSpaceCatalog(updatedSpaceCatalog)
+    finalizeSpaceSelectionChange()
+    sessionDidChange()
+  }
+
+  func isSpaceNameAvailable(
+    _ proposedName: String,
+    excluding excludedSpaceID: TerminalSpaceID? = nil
+  ) -> Bool {
+    spaceManager.isNameAvailable(proposedName, excluding: excludedSpaceID)
+  }
+
+  func observeSpaceCatalog() {
+    spaceCatalogObservationTask?.cancel()
+    spaceCatalogObservationTask = Task { @MainActor [weak self] in
+      let observations = Observations { [weak self] in
+        self?.spaceCatalog ?? .default
+      }
+      for await spaceCatalog in observations {
+        guard let self else { return }
+        self.applyObservedSpaceCatalog(spaceCatalog)
+      }
+    }
+  }
+
+  func applyObservedSpaceCatalog(_ spaceCatalog: TerminalSpaceCatalog) {
+    let resolvedSpaceCatalog = TerminalSpaceCatalog.sanitized(spaceCatalog)
+    guard resolvedSpaceCatalog != lastAppliedSpaceCatalog else { return }
+
+    let previousSelectedSpaceID = selectedSpaceID
+    lastAppliedSpaceCatalog = resolvedSpaceCatalog
+    let diff = spaceManager.applyCatalog(resolvedSpaceCatalog)
+    removeTrees(for: diff.removedTabIDs, source: .spaceCatalogObserved)
+    synchronizePinnedTabCatalogWithSpaces()
+
+    if previousSelectedSpaceID != selectedSpaceID {
+      finalizeSpaceSelectionChange()
+      sessionDidChange()
+    } else if !diff.removedTabIDs.isEmpty {
+      syncFocus(windowActivity)
+      sessionDidChange()
+    }
+  }
+
+  @discardableResult
+  func writeSpaceCatalog(
+    _ spaceCatalog: TerminalSpaceCatalog
+  ) -> TerminalSpaceManager.SpaceCatalogDiff {
+    let resolvedSpaceCatalog = TerminalSpaceCatalog.sanitized(spaceCatalog)
+    replaceSpaceCatalog(resolvedSpaceCatalog)
+    lastAppliedSpaceCatalog = resolvedSpaceCatalog
+
+    let diff = spaceManager.applyCatalog(resolvedSpaceCatalog)
+    removeTrees(for: diff.removedTabIDs, source: .spaceCatalogWrite)
+    synchronizePinnedTabCatalogWithSpaces()
+    return diff
+  }
+
+  func persistDefaultSelectedSpaceID(_ spaceID: TerminalSpaceID) {
+    guard spaceCatalog.defaultSelectedSpaceID != spaceID else { return }
+
+    var updatedSpaceCatalog = spaceCatalog
+    updatedSpaceCatalog.defaultSelectedSpaceID = spaceID
+    replaceSpaceCatalog(updatedSpaceCatalog)
+    lastAppliedSpaceCatalog = updatedSpaceCatalog
+  }
+
+  func replaceSpaceCatalog(_ spaceCatalog: TerminalSpaceCatalog) {
+    $spaceCatalog.withLock { $0 = spaceCatalog }
+  }
+}
