+++ apps/mac/supaterm/Features/Terminal/TerminalHostState.swift
@@ -529,162 +529,6 @@ final class TerminalHostState {
     }
   }
 
-  func ensureInitialTab(
-    focusing: Bool,
-    startupCommand: String? = nil,
-    workingDirectoryPath: String? = nil
-  ) {
-    guard tabs.isEmpty else { return }
-    _ = createTab(
-      focusing: focusing,
-      startupCommand: startupCommand,
-      workingDirectoryPath: workingDirectoryPath
-    )
-  }
-
-  @discardableResult
-  func createTab(
-    focusing: Bool = true,
-    startupCommand: String? = nil,
-    workingDirectoryPath: String? = nil,
-    inheritingFromSurfaceID: UUID? = nil,
-    sessionChangesEnabled: Bool = true
-  ) -> TerminalTabID? {
-    guard let target = resolveLocalCreateTabTarget(inheritingFromSurfaceID: inheritingFromSurfaceID)
-    else {
-      return nil
-    }
-    return createTab(
-      in: target.spaceID,
-      focusing: focusing,
-      startupCommand: startupCommand,
-      workingDirectory: workingDirectoryPath.map { URL(fileURLWithPath: $0, isDirectory: true) },
-      inheritingFromSurfaceID: target.inheritedSurfaceID,
-      insertion: resolvedNewTabInsertion(anchorTabID: target.anchorTabID),
-      sessionChangesEnabled: sessionChangesEnabled
-    )
-  }
-
-  @discardableResult
-  func createTab(
-    in spaceID: TerminalSpaceID,
-    focusing: Bool = true,
-    startupCommand: String? = nil,
-    workingDirectory: URL? = nil,
-    inheritingFromSurfaceID: UUID? = nil,
-    insertion: TerminalTabManager.Insertion = .end,
-    sessionChangesEnabled: Bool = true,
-    synchronizesFocus: Bool = true
-  ) -> TerminalTabID? {
-    guard let tabManager = spaceManager.tabManager(for: spaceID) else { return nil }
-    let context: ghostty_surface_context_e =
-      tabManager.tabs.isEmpty
-      ? GHOSTTY_SURFACE_CONTEXT_WINDOW
-      : GHOSTTY_SURFACE_CONTEXT_TAB
-    let tabID = tabManager.createTab(
-      title: "Terminal \(nextTabIndex(in: spaceID))",
-      icon: "terminal",
-      insertion: insertion
-    )
-    let tree = splitTree(
-      for: tabID,
-      inheritingFromSurfaceID: inheritingFromSurfaceID,
-      startupCommand: startupCommand,
-      workingDirectory: workingDirectory,
-      context: context
-    )
-    updateRunningState(for: tabID)
-    updateTabTitle(for: tabID)
-    if focusing, let surface = tree.root?.leftmostLeaf() {
-      focusSurface(surface, in: tabID)
-    }
-    if synchronizesFocus {
-      syncFocus(windowActivity)
-    }
-    if sessionChangesEnabled {
-      sessionDidChange()
-    }
-    return tabID
-  }
-
-  @discardableResult
-  func applySelectedSpace(_ spaceID: TerminalSpaceID) -> Bool {
-    let currentSelectedSpaceID = selectedSpaceID
-    guard spaceManager.selectSpace(spaceID) else { return false }
-    if currentSelectedSpaceID != spaceID, let currentSelectedSpaceID {
-      previousSelectedSpaceID = currentSelectedSpaceID
-    }
-    return true
-  }
-
-  func applySelectedTab(
-    _ tabID: TerminalTabID,
-    in spaceID: TerminalSpaceID
-  ) {
-    let currentSelectedTabID = spaceManager.selectedTabID(in: spaceID)
-    if currentSelectedTabID != tabID, let currentSelectedTabID {
-      previousSelectedTabIDBySpace[spaceID] = currentSelectedTabID
-    }
-    spaceManager.tabManager(for: spaceID)?.selectTab(tabID)
-  }
-
-  func selectTab(_ tabID: TerminalTabID) {
-    guard let space = spaceManager.space(for: tabID) else { return }
-    let didChangeSpace = spaceManager.selectedSpaceID != space.id
-    guard applySelectedSpace(space.id) else { return }
-    if didChangeSpace {
-      persistDefaultSelectedSpaceID(space.id)
-    }
-    applySelectedTab(tabID, in: space.id)
-    focusSurface(in: tabID)
-    syncFocus(windowActivity)
-    sessionDidChange()
-  }
-
-  func selectTab(slot: Int) {
-    let index = slot - 1
-    guard visibleTabs.indices.contains(index) else { return }
-    selectTab(visibleTabs[index].id)
-  }
-
-  func nextTab() {
-    guard
-      let selectedTabID,
-      let selectedIndex = tabs.firstIndex(where: { $0.id == selectedTabID }),
-      !tabs.isEmpty
-    else {
-      return
-    }
-    let nextIndex = (selectedIndex + 1) % tabs.count
-    selectTab(tabs[nextIndex].id)
-  }
-
-  func nextSpace() {
-    selectAdjacentSpace(step: 1)
-  }
-
-  func previousTab() {
-    guard
-      let selectedTabID,
-      let selectedIndex = tabs.firstIndex(where: { $0.id == selectedTabID }),
-      !tabs.isEmpty
-    else {
-      return
-    }
-    let previousIndex = (selectedIndex - 1 + tabs.count) % tabs.count
-    selectTab(tabs[previousIndex].id)
-  }
-
-  func previousSpace() {
-    selectAdjacentSpace(step: -1)
-  }
-
-  func selectLastTab() {
-    guard let selectedSpaceID else { return }
-    guard let lastTabID = previousSelectedTabIDBySpace[selectedSpaceID] else { return }
-    selectTab(lastTabID)
-  }
-
   func setPinnedTabOrder(_ orderedIDs: [TerminalTabID]) {
     spaceManager.activeTabManager?.setPinnedTabOrder(orderedIDs)
     if let selectedSpaceID {
@@ -721,165 +565,6 @@ final class TerminalHostState {
     }
     syncPinnedTabMembership(in: spaceID)
     sessionDidChange()
-  }
-
-  @discardableResult
-  func createSpace(named name: String) throws -> TerminalSpaceID {
-    guard let normalizedName = Self.trimmedNonEmpty(name) else {
-      throw TerminalControlError.invalidSpaceName
-    }
-    guard spaceManager.isNameAvailable(normalizedName) else {
-      throw TerminalControlError.spaceNameUnavailable
-    }
-
-    let space = PersistedTerminalSpace(name: normalizedName)
-    var updatedSpaceCatalog = spaceCatalog
-    updatedSpaceCatalog.defaultSelectedSpaceID = space.id
-    updatedSpaceCatalog = TerminalSpaceCatalog(
-      defaultSelectedSpaceID: updatedSpaceCatalog.defaultSelectedSpaceID,
-      spaces: updatedSpaceCatalog.spaces + [space]
-    )
-    _ = writeSpaceCatalog(updatedSpaceCatalog)
-    guard applySelectedSpace(space.id) else {
-      throw TerminalControlError.spaceNotFound(windowIndex: 1, spaceIndex: spaces.count + 1)
-    }
-    finalizeSpaceSelectionChange()
-    sessionDidChange()
-    return space.id
-  }
-
-  func selectSpace(_ spaceID: TerminalSpaceID) {
-    selectSpace(spaceID, persistDefaultSelection: true)
-  }
-
-  func selectSpace(
-    _ spaceID: TerminalSpaceID,
-    persistDefaultSelection: Bool
-  ) {
-    guard applySelectedSpace(spaceID) else { return }
-    if persistDefaultSelection {
-      persistDefaultSelectedSpaceID(spaceID)
-    }
-    finalizeSpaceSelectionChange()
-    sessionDidChange()
-  }
-
-  func selectSpace(slot: Int) {
-    let index = slot == 0 ? 9 : slot - 1
-    guard spaces.indices.contains(index) else { return }
-    selectSpace(spaces[index].id)
-  }
-
-  func selectAdjacentSpace(step: Int) {
-    guard
-      spaces.count > 1,
-      let selectedSpaceID,
-      let currentIndex = spaces.firstIndex(where: { $0.id == selectedSpaceID })
-    else { return }
-
-    let targetIndex = (currentIndex + step + spaces.count) % spaces.count
-    selectSpace(spaces[targetIndex].id)
-  }
-
-  func renameSpace(_ spaceID: TerminalSpaceID, to name: String) {
-    let normalizedName = name.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !normalizedName.isEmpty else { return }
-    guard spaceManager.isNameAvailable(normalizedName, excluding: spaceID) else { return }
-    guard let index = spaceCatalog.spaces.firstIndex(where: { $0.id == spaceID }) else { return }
-
-    var updatedSpaceCatalog = spaceCatalog
-    updatedSpaceCatalog.spaces[index].name = normalizedName
-    _ = writeSpaceCatalog(updatedSpaceCatalog)
-  }
-
-  func deleteSpace(_ spaceID: TerminalSpaceID) {
-    let remainingSpaces = spaceCatalog.spaces.filter { $0.id != spaceID }
-    guard !remainingSpaces.isEmpty else { return }
-    guard remainingSpaces.count != spaceCatalog.spaces.count else { return }
-    if previousSelectedSpaceID == spaceID {
-      previousSelectedSpaceID = nil
-    }
-
-    let nextSelectedSpaceID = nextSelectedSpaceID(afterDeleting: spaceID)
-    let updatedSpaceCatalog = TerminalSpaceCatalog(
-      defaultSelectedSpaceID: nextSelectedSpaceID,
-      spaces: remainingSpaces
-    )
-    _ = writeSpaceCatalog(updatedSpaceCatalog)
-    finalizeSpaceSelectionChange()
-    sessionDidChange()
-  }
-
-  func isSpaceNameAvailable(
-    _ proposedName: String,
-    excluding excludedSpaceID: TerminalSpaceID? = nil
-  ) -> Bool {
-    spaceManager.isNameAvailable(proposedName, excluding: excludedSpaceID)
-  }
-
-  func closeSurface(_ surfaceID: UUID) {
-    performCloseSurface(surfaceID, source: .commandCloseSurface)
-  }
-
-  func closeTab(_ tabID: TerminalTabID) {
-    performCloseTab(tabID)
-  }
-
-  func closeTabs(_ tabIDs: [TerminalTabID]) {
-    performCloseTabs(tabIDs)
-  }
-
-  func requestCloseSurface(
-    _ surfaceID: UUID,
-    needsConfirmation: Bool? = nil,
-    source: TerminalSurfaceCloseSource = .commandRequestCloseSurface
-  ) {
-    guard
-      let resolvedCloseRequest = resolvedCloseRequest(
-        for: .surface(surfaceID),
-        needsConfirmationOverride: needsConfirmation
-      )
-    else {
-      SupatermLog.debug(
-        SupatermLog.terminal,
-        "terminal.close.request.dropped",
-        fields: [
-          "source=\(source.rawValue)",
-          "surfaceID=\(SupatermLog.uuid(surfaceID))",
-          "reason=unresolved",
-        ]
-      )
-      return
-    }
-    logCloseRequest(
-      surfaceID: surfaceID,
-      needsConfirmationOverride: needsConfirmation,
-      resolvedCloseRequest: resolvedCloseRequest,
-      source: source
-    )
-    emit(resolvedCloseRequest)
-  }
-
-  func requestCloseTab(_ tabID: TerminalTabID) {
-    guard let resolvedCloseRequest = resolvedCloseRequest(for: .tab(tabID)) else { return }
-    emit(resolvedCloseRequest)
-  }
-
-  func requestCloseTabsBelow(_ tabID: TerminalTabID) {
-    guard let space = spaceManager.space(for: tabID) else { return }
-    guard let tabManager = spaceManager.tabManager(for: space.id) else { return }
-    requestCloseTabs(tabManager.tabIDsBelow(tabID))
-  }
-
-  func requestCloseOtherTabs(_ tabID: TerminalTabID) {
-    guard let space = spaceManager.space(for: tabID) else { return }
-    guard let tabManager = spaceManager.tabManager(for: space.id) else { return }
-    requestCloseTabs(tabManager.otherTabIDs(tabID))
-  }
-
-  func requestCloseTabs(_ tabIDs: [TerminalTabID]) {
-    guard let resolvedCloseRequest = resolvedCloseRequest(for: .tabs(tabIDs)) else { return }
-    emit(resolvedCloseRequest)
   }
 
   @discardableResult
@@ -1161,196 +846,6 @@ final class TerminalHostState {
     }
   }
 
-  func performCloseSurface(
-    _ surfaceID: UUID,
-    source: TerminalSurfaceCloseSource = .commandCloseSurface
-  ) {
-    guard let tabID = tabID(containing: surfaceID), let tree = trees[tabID] else {
-      logClosePerformDropped(surfaceID: surfaceID, source: source, reason: "missingTree")
-      return
-    }
-    guard let node = tree.find(id: surfaceID), let surface = surfaces[surfaceID] else {
-      logClosePerformDropped(
-        surfaceID: surfaceID,
-        tabID: tabID,
-        source: source,
-        reason: "missingSurface"
-      )
-      return
-    }
-    let spaceID = spaceManager.space(for: tabID)?.id
-    let wasPinned = spaceManager.tab(for: tabID)?.isPinned == true
-    let wasSelectedSpace = selectedSpaceID == spaceID
-
-    let nextSurface =
-      focusHistoryByTab[tabID]?.current == surfaceID
-      ? tree.focusTargetAfterClosing(node)
-      : nil
-    let newTree = tree.removing(node)
-    logClosePerform(
-      TerminalClosePerformLogContext(
-        source: source,
-        surfaceID: surfaceID,
-        tabID: tabID,
-        spaceID: spaceID,
-        wasPinned: wasPinned,
-        leafCount: tree.leaves().count,
-        newTreeEmpty: newTree.isEmpty,
-        focusedSurfaceID: focusHistoryByTab[tabID]?.current,
-        nextSurfaceID: nextSurface?.id
-      )
-    )
-
-    if newTree.isEmpty, wasPinned {
-      logCloseKillSurface(surfaceID: surfaceID, tabID: tabID, source: source)
-      persistPinnedTabWorkingDirectoriesIfNeeded(for: tabID)
-      removeTree(for: tabID, source: .pinnedLastPaneClose)
-      if let spaceID {
-        spaceManager.tabManager(for: spaceID)?.updateDirty(tabID, isDirty: false)
-        updateSelectionAfterClosingTab(in: spaceID, wasSelectedSpace: wasSelectedSpace)
-      } else {
-        lastEmittedFocusSurfaceID = nil
-      }
-      syncFocus(windowActivity)
-      sessionDidChange(persistingPinnedTabLayouts: false)
-      return
-    }
-
-    logCloseKillSurface(surfaceID: surfaceID, tabID: tabID, source: source)
-    killZmxSession(for: surfaceID)
-    cleanupSurface(surface)
-
-    if newTree.isEmpty {
-      trees.removeValue(forKey: tabID)
-      focusHistoryByTab.removeValue(forKey: tabID)
-      spaceManager.space(for: tabID)
-        .flatMap { spaceManager.tabManager(for: $0.id) }?
-        .closeTab(tabID)
-      if let spaceID {
-        updateSelectionAfterClosingTab(in: spaceID, wasSelectedSpace: wasSelectedSpace)
-      } else {
-        lastEmittedFocusSurfaceID = nil
-      }
-      syncFocus(windowActivity)
-      sessionDidChange()
-      return
-    }
-
-    trees[tabID] = newTree
-    updateRunningState(for: tabID)
-    updateTabTitle(for: tabID)
-    if focusHistoryByTab[tabID]?.current == surfaceID {
-      if let nextSurface {
-        focusSurface(nextSurface, in: tabID)
-      } else {
-        focusHistoryByTab.removeValue(forKey: tabID)
-      }
-    }
-    syncFocus(windowActivity)
-    sessionDidChange(persistingPinnedTabLayouts: false)
-  }
-
-  func createSurface(
-    tabID: TerminalTabID,
-    startupCommand: String?,
-    inheritingFromSurfaceID: UUID?,
-    workingDirectory: URL? = nil,
-    context: ghostty_surface_context_e,
-    surfaceID: UUID = UUID()
-  ) -> GhosttySurfaceView {
-    guard let runtime else {
-      preconditionFailure("TerminalHostState cannot create surfaces without a GhosttyRuntime")
-    }
-    let inherited = inheritedSurfaceConfig(fromSurfaceID: inheritingFromSurfaceID, context: context)
-    let launchCommand = resolvedSurfaceCommand(
-      startupCommand: startupCommand,
-      surfaceID: surfaceID
-    )
-    SupatermLog.debug(
-      SupatermLog.terminal,
-      "terminal.surface.create",
-      fields: [
-        "surfaceID=\(surfaceID.uuidString.lowercased())",
-        "tabID=\(tabID.rawValue.uuidString.lowercased())",
-        "context=\(Self.surfaceContextLabel(context))",
-        "zmxSessionsEnabled=\(zmxSessionsEnabled)",
-        "hasStartupCommand=\(startupCommand?.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty == false)",
-        "hasResolvedCommand=\(launchCommand.command != nil)",
-        "hasCommandWrapper=\(!launchCommand.commandWrapper.isEmpty)",
-        "usesZmx=\(launchCommand.usesZmx)",
-      ]
-    )
-    let view = GhosttySurfaceView(
-      id: surfaceID,
-      runtime: runtime,
-      tabID: tabID.rawValue,
-      workingDirectory: workingDirectory ?? inherited.workingDirectory,
-      command: launchCommand.command,
-      commandWrapper: launchCommand.commandWrapper,
-      fontSize: inherited.fontSize,
-      context: context,
-      managesWindowAppearance: false,
-      zmxSessionsEnabled: launchCommand.usesZmx
-    )
-    configureBridgeCallbacks(for: view, tabID: tabID)
-    configureSurfaceCallbacks(for: view, tabID: tabID)
-    surfaces[view.id] = view
-    return view
-  }
-
-  func resolvedSurfaceCommand(
-    startupCommand: String?,
-    surfaceID: UUID
-  ) -> SurfaceLaunchCommand {
-    let command = startupCommand.map { SupatermShellCommand.ghosttyStartupCommand(for: $0) }
-    let sessionID = ZmxSessionID.make(surfaceID: surfaceID)
-    guard zmxSessionsEnabled else {
-      SupatermLog.debug(
-        SupatermLog.zmx,
-        "zmx.attach.skipped",
-        fields: [
-          "surfaceID=\(surfaceID.uuidString.lowercased())",
-          "sessionID=\(sessionID)",
-          "reason=disabled",
-        ]
-      )
-      return SurfaceLaunchCommand(command: command, commandWrapper: [], usesZmx: false)
-    }
-    guard let executable = zmxClient.executableURL() else {
-      SupatermLog.error(
-        SupatermLog.zmx,
-        "zmx.attach.fallback",
-        fields: [
-          "surfaceID=\(surfaceID.uuidString.lowercased())",
-          "sessionID=\(sessionID)",
-          "hasStartupCommand=\(command != nil)",
-        ]
-      )
-      return SurfaceLaunchCommand(command: command, commandWrapper: [], usesZmx: false)
-    }
-    let zmxCommand = startupCommand.map { SupatermShellCommand.ghosttyStartupCommand(for: $0) }
-    let launch = ZmxAttach.resolveLaunch(
-      executablePath: executable.path(percentEncoded: false),
-      sessionID: sessionID,
-      command: zmxCommand
-    )
-    SupatermLog.debug(
-      SupatermLog.zmx,
-      "zmx.attach.resolved",
-      fields: [
-        "surfaceID=\(surfaceID.uuidString.lowercased())",
-        "sessionID=\(sessionID)",
-        "hasStartupCommand=\(launch.command != nil)",
-        "hasCommandWrapper=\(!launch.commandWrapper.isEmpty)",
-      ]
-    )
-    return SurfaceLaunchCommand(
-      command: launch.command,
-      commandWrapper: launch.commandWrapper,
-      usesZmx: true
-    )
-  }
-
   static func surfaceContextLabel(_ context: ghostty_surface_context_e) -> String {
     switch context {
     case GHOSTTY_SURFACE_CONTEXT_WINDOW:
@@ -1362,102 +857,6 @@ final class TerminalHostState {
     default:
       return String(Int(context.rawValue))
     }
-  }
-
-  func suspendPinnedTab(_ tabID: TerminalTabID) {
-    guard
-      let space = spaceManager.space(for: tabID),
-      let tabManager = spaceManager.tabManager(for: space.id),
-      spaceManager.tab(for: tabID)?.isPinned == true
-    else {
-      return
-    }
-    let surfaceIDs = trees[tabID]?.leaves().map(\.id) ?? []
-    SupatermLog.debug(
-      SupatermLog.terminal,
-      "terminal.pinned.suspend",
-      fields: [
-        "tabID=\(SupatermLog.uuid(tabID.rawValue))",
-        "spaceID=\(SupatermLog.uuid(space.id.rawValue))",
-        "surfaceIDs=\(Self.logSurfaceIDs(surfaceIDs))",
-      ]
-    )
-    persistPinnedTabWorkingDirectoriesIfNeeded(for: tabID)
-    removeTree(for: tabID, terminateSessions: false, source: .pinnedSuspend)
-    tabManager.updateDirty(tabID, isDirty: false)
-  }
-
-  func updateSelectionAfterClosingTab(
-    in spaceID: TerminalSpaceID,
-    wasSelectedSpace: Bool
-  ) {
-    guard wasSelectedSpace else { return }
-
-    if let selectedTabID = spaceManager.selectedTabID(in: spaceID) {
-      if isSelectableTab(selectedTabID) {
-        focusSurfaceIfNeeded(in: selectedTabID)
-        return
-      }
-    }
-
-    if let tabID = replacementLiveTabID(in: spaceID) {
-      applySelectedTab(tabID, in: spaceID)
-      focusSurfaceIfNeeded(in: tabID)
-      return
-    }
-
-    spaceManager.tabManager(for: spaceID)?.clearSelection()
-
-    if let previousSelectedSpaceID,
-      previousSelectedSpaceID != spaceID,
-      let tabID = replacementLiveTabID(in: previousSelectedSpaceID)
-    {
-      _ = applySelectedSpace(previousSelectedSpaceID)
-      applySelectedTab(tabID, in: previousSelectedSpaceID)
-      focusSurfaceIfNeeded(in: tabID)
-      return
-    }
-
-    if let fallback = firstLiveTabLocation() {
-      _ = applySelectedSpace(fallback.spaceID)
-      applySelectedTab(fallback.tabID, in: fallback.spaceID)
-      focusSurfaceIfNeeded(in: fallback.tabID)
-      return
-    }
-
-    lastEmittedFocusSurfaceID = nil
-  }
-
-  func replacementLiveTabID(in spaceID: TerminalSpaceID) -> TerminalTabID? {
-    let tabs = spaceManager.tabs(in: spaceID)
-    if let previousTabID = previousSelectedTabIDBySpace[spaceID],
-      tabs.contains(where: { $0.id == previousTabID }),
-      isSelectableTab(previousTabID)
-    {
-      return previousTabID
-    }
-    return tabs.reversed().first { isSelectableTab($0.id) }?.id
-  }
-
-  func firstLiveTabLocation() -> (spaceID: TerminalSpaceID, tabID: TerminalTabID)? {
-    for space in spaces {
-      if let tabID = spaceManager.tabs(in: space.id).first(where: { isSelectableTab($0.id) })?.id {
-        return (space.id, tabID)
-      }
-    }
-    return nil
-  }
-
-  func isSelectableTab(_ tabID: TerminalTabID) -> Bool {
-    !managesTerminalSurfaces || trees[tabID] != nil
-  }
-
-  func focusSurfaceIfNeeded(in tabID: TerminalTabID) {
-    guard managesTerminalSurfaces else {
-      lastEmittedFocusSurfaceID = nil
-      return
-    }
-    focusSurface(in: tabID)
   }
 
   func configureBridgeCallbacks(
@@ -1573,177 +972,6 @@ final class TerminalHostState {
     }
   }
 
-  func requestCloseSurfaceAfterProcessExit(
-    _ surfaceID: UUID,
-    source: TerminalSurfaceCloseSource
-  ) {
-    guard zmxSessionsEnabled, zmxClient.isBundled() else {
-      requestCloseSurface(
-        surfaceID,
-        needsConfirmation: false,
-        source: source
-      )
-      return
-    }
-
-    let zmxClient = zmxClient
-    let sessionID = ZmxSessionID.make(surfaceID: surfaceID)
-    Task { @MainActor [weak self, zmxClient] in
-      let sessionIDs = await zmxClient.listSessions()
-      guard let self else { return }
-      self.finishCloseSurfaceAfterProcessExit(
-        surfaceID,
-        sessionIDs: sessionIDs,
-        sessionID: sessionID,
-        source: source,
-        didRetry: false
-      )
-    }
-  }
-
-  func finishCloseSurfaceAfterProcessExit(
-    _ surfaceID: UUID,
-    sessionIDs: [String]?,
-    sessionID: String,
-    source: TerminalSurfaceCloseSource,
-    didRetry: Bool
-  ) {
-    guard let sessionIDs else {
-      guard !didRetry else {
-        requestCloseSurface(
-          surfaceID,
-          needsConfirmation: false,
-          source: source
-        )
-        return
-      }
-      let zmxClient = zmxClient
-      Task { @MainActor [weak self, zmxClient] in
-        try? await Task.sleep(for: .milliseconds(150))
-        let retrySessionIDs = await zmxClient.listSessions()
-        guard let self else { return }
-        self.finishCloseSurfaceAfterProcessExit(
-          surfaceID,
-          sessionIDs: retrySessionIDs,
-          sessionID: sessionID,
-          source: source,
-          didRetry: true
-        )
-      }
-      return
-    }
-    guard sessionIDs.contains(sessionID), reattachZmxSurface(surfaceID, source: source) else {
-      requestCloseSurface(
-        surfaceID,
-        needsConfirmation: false,
-        source: source
-      )
-      return
-    }
-  }
-
-  @discardableResult
-  func reattachZmxSurface(
-    _ surfaceID: UUID,
-    source: TerminalSurfaceCloseSource
-  ) -> Bool {
-    guard let tabID = tabID(containing: surfaceID), var tree = trees[tabID] else {
-      SupatermLog.debug(
-        SupatermLog.terminal,
-        "terminal.close.zmxReattach.dropped",
-        fields: [
-          "source=\(source.rawValue)",
-          "surfaceID=\(SupatermLog.uuid(surfaceID))",
-          "reason=missingTree",
-        ]
-      )
-      return false
-    }
-    guard let node = tree.find(id: surfaceID), let previousSurface = surfaces[surfaceID] else {
-      SupatermLog.debug(
-        SupatermLog.terminal,
-        "terminal.close.zmxReattach.dropped",
-        fields: [
-          "source=\(source.rawValue)",
-          "surfaceID=\(SupatermLog.uuid(surfaceID))",
-          "tabID=\(SupatermLog.uuid(tabID.rawValue))",
-          "reason=missingSurface",
-        ]
-      )
-      return false
-    }
-
-    let context = reattachSurfaceContext(for: tabID, tree: tree)
-    let workingDirectory = existingWorkingDirectoryURL(for: workingDirectoryPath(for: previousSurface))
-    let titleOverride = previousSurface.bridge.state.titleOverride
-    previousSurface.bridge.onChildExited = nil
-    previousSurface.bridge.onCloseRequest = nil
-
-    let replacementSurface = createSurface(
-      tabID: tabID,
-      startupCommand: nil,
-      inheritingFromSurfaceID: nil,
-      workingDirectory: workingDirectory,
-      context: context,
-      surfaceID: surfaceID
-    )
-    replacementSurface.bridge.state.titleOverride = titleOverride
-
-    do {
-      tree = try tree.replacing(node: node, with: .leaf(view: replacementSurface))
-    } catch {
-      surfaces.removeValue(forKey: surfaceID)
-      replacementSurface.closeSurface()
-      configureBridgeCloseCallbacks(for: previousSurface)
-      surfaces[surfaceID] = previousSurface
-      SupatermLog.error(
-        SupatermLog.terminal,
-        "terminal.close.zmxReattach.failed",
-        fields: [
-          "source=\(source.rawValue)",
-          "surfaceID=\(SupatermLog.uuid(surfaceID))",
-          "tabID=\(SupatermLog.uuid(tabID.rawValue))",
-          "error=\(String(describing: error))",
-        ]
-      )
-      return false
-    }
-
-    previousSurface.closeSurface()
-    trees[tabID] = tree
-    updateRunningState(for: tabID)
-    updateTabTitle(for: tabID)
-    if focusHistoryByTab[tabID]?.current == surfaceID {
-      focusSurface(replacementSurface, in: tabID)
-    }
-    syncFocus(windowActivity)
-    SupatermLog.debug(
-      SupatermLog.terminal,
-      "terminal.close.zmxReattach.finished",
-      fields: [
-        "source=\(source.rawValue)",
-        "surfaceID=\(SupatermLog.uuid(surfaceID))",
-        "tabID=\(SupatermLog.uuid(tabID.rawValue))",
-        "context=\(Self.surfaceContextLabel(context))",
-      ]
-    )
-    return true
-  }
-
-  func reattachSurfaceContext(
-    for tabID: TerminalTabID,
-    tree: SplitTree<GhosttySurfaceView>
-  ) -> ghostty_surface_context_e {
-    guard !tree.isSplit else { return GHOSTTY_SURFACE_CONTEXT_SPLIT }
-    guard
-      let spaceID = spaceManager.space(for: tabID)?.id,
-      spaceManager.tabs(in: spaceID).first?.id == tabID
-    else {
-      return GHOSTTY_SURFACE_CONTEXT_TAB
-    }
-    return GHOSTTY_SURFACE_CONTEXT_WINDOW
-  }
-
   func configureSurfaceCallbacks(
     for view: GhosttySurfaceView,
     tabID: TerminalTabID
@@ -1767,57 +995,6 @@ final class TerminalHostState {
   struct InheritedSurfaceConfig: Equatable {
     let workingDirectory: URL?
     let fontSize: Float32?
-  }
-
-  func inheritedSurfaceConfig(
-    fromSurfaceID surfaceID: UUID?,
-    context: ghostty_surface_context_e
-  ) -> InheritedSurfaceConfig {
-    guard let surfaceID, let view = surfaces[surfaceID], let sourceSurface = view.surface else {
-      return InheritedSurfaceConfig(workingDirectory: nil, fontSize: nil)
-    }
-
-    let inherited = ghostty_surface_inherited_config(sourceSurface, context)
-    let fontSize = inherited.font_size == 0 ? nil : inherited.font_size
-    let workingDirectory = inherited.working_directory.flatMap { ptr -> URL? in
-      let path = String(cString: ptr)
-      guard !path.isEmpty else { return nil }
-      return URL(fileURLWithPath: path, isDirectory: true)
-    }
-
-    return InheritedSurfaceConfig(
-      workingDirectory: workingDirectory,
-      fontSize: fontSize
-    )
-  }
-
-  func currentFocusedSurfaceID() -> UUID? {
-    guard let selectedTabID = spaceManager.selectedTabID else { return nil }
-    return focusHistoryByTab[selectedTabID]?.current
-  }
-
-  func inheritedSurfaceID(in spaceID: TerminalSpaceID) -> UUID? {
-    if let selectedTabID = spaceManager.selectedTabID(in: spaceID) {
-      if let focusedSurfaceID = focusHistoryByTab[selectedTabID]?.current,
-        surfaces[focusedSurfaceID] != nil
-      {
-        return focusedSurfaceID
-      }
-      if let surfaceID = trees[selectedTabID]?.root?.leftmostLeaf().id {
-        return surfaceID
-      }
-    }
-
-    for tab in spaceManager.tabs(in: spaceID) {
-      if let focusedSurfaceID = focusHistoryByTab[tab.id]?.current, surfaces[focusedSurfaceID] != nil {
-        return focusedSurfaceID
-      }
-      if let surfaceID = trees[tab.id]?.root?.leftmostLeaf().id {
-        return surfaceID
-      }
-    }
-
-    return nil
   }
 
   struct ResolvedCreateTabTarget {
@@ -1850,143 +1027,10 @@ final class TerminalHostState {
     let tree: SplitTree<GhosttySurfaceView>
   }
 
-  func resolveCreateTabTarget(
-    _ target: TerminalCreateTabRequest.Target
-  ) throws -> ResolvedCreateTabTarget {
-    switch target {
-    case .contextPane(let paneID):
-      guard
-        let tabID = tabID(containing: paneID),
-        let space = spaceManager.space(for: tabID)
-      else {
-        throw TerminalCreateTabError.contextPaneNotFound
-      }
-
-      return ResolvedCreateTabTarget(
-        anchorTabID: tabID,
-        inheritedSurfaceID: paneID,
-        space: space
-      )
-
-    case .space(let windowIndex, let spaceIndex):
-      guard windowIndex == 1 else {
-        throw TerminalCreateTabError.windowNotFound(windowIndex)
-      }
-      guard let space = spaceManager.space(at: spaceIndex) else {
-        throw TerminalCreateTabError.spaceNotFound(windowIndex: windowIndex, spaceIndex: spaceIndex)
-      }
-      return ResolvedCreateTabTarget(
-        anchorTabID: spaceManager.selectedTabID(in: space.id),
-        inheritedSurfaceID: inheritedSurfaceID(in: space.id),
-        space: space
-      )
-    }
-  }
-
-  func resolveLocalCreateTabTarget(
-    inheritingFromSurfaceID: UUID?
-  ) -> ResolvedLocalCreateTabTarget? {
-    if let inheritingFromSurfaceID,
-      let anchorTabID = tabID(containing: inheritingFromSurfaceID),
-      let space = spaceManager.space(for: anchorTabID)
-    {
-      return ResolvedLocalCreateTabTarget(
-        anchorTabID: anchorTabID,
-        inheritedSurfaceID: inheritingFromSurfaceID,
-        spaceID: space.id
-      )
-    }
-
-    guard let spaceID = spaceManager.selectedSpaceID else {
-      return nil
-    }
-
-    return ResolvedLocalCreateTabTarget(
-      anchorTabID: spaceManager.selectedTabID(in: spaceID),
-      inheritedSurfaceID: inheritingFromSurfaceID ?? currentFocusedSurfaceID(),
-      spaceID: spaceID
-    )
-  }
-
-  func resolvedNewTabInsertion(
-    anchorTabID: TerminalTabID?
-  ) -> TerminalTabManager.Insertion {
-    switch supatermSettings.newTabPosition {
-    case .current:
-      guard let anchorTabID else {
-        return .end
-      }
-      return .after(anchorTabID)
-    case .end:
-      return .end
-    }
-  }
-
   struct ResolvedPaneLocation {
     let paneIndex: Int
     let spaceIndex: Int
     let tabIndex: Int
-  }
-
-  func resolveCreatePaneTarget(
-    _ target: TerminalCreatePaneRequest.Target
-  ) throws -> ResolvedCreatePaneTarget {
-    switch target {
-    case .contextPane(let paneID):
-      guard
-        let tabID = tabID(containing: paneID),
-        let space = spaceManager.space(for: tabID),
-        let tree = trees[tabID],
-        let anchorSurface = surfaces[paneID]
-      else {
-        throw TerminalCreatePaneError.contextPaneNotFound
-      }
-
-      return ResolvedCreatePaneTarget(
-        anchorSurface: anchorSurface,
-        spaceID: space.id,
-        tabID: tabID,
-        tree: tree
-      )
-
-    case .pane(let windowIndex, let spaceIndex, let tabIndex, let paneIndex):
-      let resolvedTab = try resolveTab(
-        windowIndex: windowIndex, spaceIndex: spaceIndex, tabIndex: tabIndex)
-      let panes = resolvedTab.tree.leaves()
-      let paneOffset = paneIndex - 1
-      guard panes.indices.contains(paneOffset) else {
-        throw TerminalCreatePaneError.paneNotFound(
-          windowIndex: windowIndex,
-          spaceIndex: spaceIndex,
-          tabIndex: tabIndex,
-          paneIndex: paneIndex
-        )
-      }
-
-      return ResolvedCreatePaneTarget(
-        anchorSurface: panes[paneOffset],
-        spaceID: resolvedTab.space.id,
-        tabID: resolvedTab.tabID,
-        tree: resolvedTab.tree
-      )
-
-    case .tab(let windowIndex, let spaceIndex, let tabIndex):
-      let resolvedTab = try resolveTab(
-        windowIndex: windowIndex, spaceIndex: spaceIndex, tabIndex: tabIndex)
-      let anchorSurface =
-        focusHistoryByTab[resolvedTab.tabID].flatMap { surfaces[$0.current] }
-        ?? resolvedTab.tree.root?.leftmostLeaf()
-      guard let anchorSurface else {
-        throw TerminalCreatePaneError.creationFailed
-      }
-
-      return ResolvedCreatePaneTarget(
-        anchorSurface: anchorSurface,
-        spaceID: resolvedTab.space.id,
-        tabID: resolvedTab.tabID,
-        tree: resolvedTab.tree
-      )
-    }
   }
 
   func resolveNotifyTarget(
@@ -2175,43 +1219,6 @@ final class TerminalHostState {
     )
   }
 
-  private func cleanupSurface(_ surface: GhosttySurfaceView) {
-    agentPanelController?.surfaceRemoved(surface.id)
-    notificationStore.removeSurface(surface.id)
-    paneAgentMetadataBySurfaceID.removeValue(forKey: surface.id)
-    agentPresenceStore.removeSurface(surface.id)
-    surface.closeSurface()
-    surfaces.removeValue(forKey: surface.id)
-  }
-
-  func removeTree(
-    for tabID: TerminalTabID,
-    terminateSessions: Bool = true,
-    source: TerminalTreeRemovalSource
-  ) {
-    guard let tree = trees.removeValue(forKey: tabID) else { return }
-    let surfaceIDs = tree.leaves().map(\.id)
-    SupatermLog.debug(
-      SupatermLog.terminal,
-      "terminal.tree.remove",
-      fields: [
-        "source=\(source.rawValue)",
-        "tabID=\(SupatermLog.uuid(tabID.rawValue))",
-        "isPinned=\(spaceManager.tab(for: tabID)?.isPinned == true)",
-        "terminateSessions=\(terminateSessions)",
-        "surfaceIDs=\(Self.logSurfaceIDs(surfaceIDs))",
-      ]
-    )
-    if terminateSessions {
-      killZmxSessions(for: surfaceIDs)
-    }
-    for surface in tree.leaves() {
-      cleanupSurface(surface)
-    }
-    focusHistoryByTab.removeValue(forKey: tabID)
-    previousSelectedTabIDBySpace = previousSelectedTabIDBySpace.filter { $0.value != tabID }
-  }
-
   func notifications(for tabID: TerminalTabID) -> [UUID: [PaneNotification]] {
     guard let tree = trees[tabID] else { return [:] }
     return Dictionary(
@@ -2283,197 +1290,8 @@ final class TerminalHostState {
       .updateDirty(tabID, isDirty: isRunning)
   }
 
-  func tabNeedsCloseConfirmation(_ tabID: TerminalTabID) -> Bool {
-    guard let tree = trees[tabID] else { return false }
-    return tree.leaves().contains(where: \.needsCloseConfirmation)
-  }
-
-  func shouldCloseWindow<S: Sequence>(afterClosing tabIDs: S) -> Bool where S.Element == TerminalTabID {
-    let requestedTabIDs = Set(tabIDs)
-    let closingRegularTabIDs = Set(
-      requestedTabIDs.filter { tabID in
-        trees[tabID] != nil && spaceManager.tab(for: tabID)?.isPinned != true
-      }
-    )
-    let survivingTabs = spaces.flatMap { space in
-      spaceManager.tabs(in: space.id).filter { tab in
-        tab.isPinned || !requestedTabIDs.contains(tab.id)
-      }
-    }
-    return !closingRegularTabIDs.isEmpty && survivingTabs.isEmpty
-  }
-
-  func shouldCloseWindow(afterClosingSurface surfaceID: UUID) -> Bool {
-    guard
-      let tabID = tabID(containing: surfaceID),
-      let tree = trees[tabID],
-      let node = tree.find(id: surfaceID)
-    else {
-      return false
-    }
-    guard tree.removing(node).isEmpty else { return false }
-    return shouldCloseWindow(afterClosing: [tabID])
-  }
-
-  func windowNeedsCloseConfirmation() -> Bool {
-    if let runtime {
-      return runtime.needsConfirmQuit()
-    }
-    return trees.values.contains { tree in
-      tree.leaves().contains(where: \.needsCloseConfirmation)
-    }
-  }
-
-  func surfaceNeedsCloseConfirmation(_ surfaceID: UUID) -> Bool {
-    surfaces[surfaceID]?.needsCloseConfirmation ?? false
-  }
-
-  func resolvedCloseRequest(
-    for target: TerminalCloseTarget,
-    needsConfirmationOverride: Bool? = nil
-  ) -> ResolvedCloseRequest? {
-    switch target {
-    case .surface(let surfaceID):
-      guard surfaces[surfaceID] != nil else { return nil }
-      if shouldCloseWindow(afterClosingSurface: surfaceID) {
-        return .window(
-          needsConfirmation: needsConfirmationOverride ?? windowNeedsCloseConfirmation()
-        )
-      }
-      return .request(
-        TerminalCloseRequest(
-          target: .surface(surfaceID),
-          needsConfirmation: needsConfirmationOverride ?? surfaceNeedsCloseConfirmation(surfaceID)
-        )
-      )
-
-    case .tab(let tabID):
-      guard trees[tabID] != nil else { return nil }
-      if shouldCloseWindow(afterClosing: [tabID]) {
-        return .window(needsConfirmation: needsConfirmationOverride ?? windowNeedsCloseConfirmation())
-      }
-      return .request(
-        TerminalCloseRequest(
-          target: .tab(tabID),
-          needsConfirmation: needsConfirmationOverride ?? tabNeedsCloseConfirmation(tabID)
-        )
-      )
-
-    case .tabs(let tabIDs):
-      let existingTabIDs = tabIDs.filter { spaceManager.tab(for: $0) != nil }
-      guard !existingTabIDs.isEmpty else { return nil }
-      if shouldCloseWindow(afterClosing: existingTabIDs) {
-        return .window(needsConfirmation: needsConfirmationOverride ?? windowNeedsCloseConfirmation())
-      }
-      let tabIDsWithTrees = existingTabIDs.filter { trees[$0] != nil }
-      return .request(
-        TerminalCloseRequest(
-          target: .tabs(existingTabIDs),
-          needsConfirmation: needsConfirmationOverride
-            ?? tabIDsWithTrees.contains(where: tabNeedsCloseConfirmation)
-        )
-      )
-    }
-  }
-
-  func tabID(containing surfaceID: UUID) -> TerminalTabID? {
-    for (tabID, tree) in trees where tree.find(id: surfaceID) != nil {
-      return tabID
-    }
-    return nil
-  }
-
   nonisolated static func logSurfaceIDs(_ surfaceIDs: some Sequence<UUID>) -> String {
     surfaceIDs.map { SupatermLog.uuid($0) }.sorted().joined(separator: ",")
-  }
-
-  func logClosePerformDropped(
-    surfaceID: UUID,
-    tabID: TerminalTabID? = nil,
-    source: TerminalSurfaceCloseSource,
-    reason: String
-  ) {
-    var fields = [
-      "source=\(source.rawValue)",
-      "surfaceID=\(SupatermLog.uuid(surfaceID))",
-    ]
-    if let tabID {
-      fields.append("tabID=\(SupatermLog.uuid(tabID.rawValue))")
-    }
-    fields.append("reason=\(reason)")
-    SupatermLog.debug(
-      SupatermLog.terminal,
-      "terminal.close.perform.dropped",
-      fields: fields
-    )
-  }
-
-  func logClosePerform(_ context: TerminalClosePerformLogContext) {
-    SupatermLog.debug(
-      SupatermLog.terminal,
-      "terminal.close.perform",
-      fields: [
-        "source=\(context.source.rawValue)",
-        "surfaceID=\(SupatermLog.uuid(context.surfaceID))",
-        "tabID=\(SupatermLog.uuid(context.tabID.rawValue))",
-        "spaceID=\(SupatermLog.uuid(context.spaceID?.rawValue))",
-        "wasPinned=\(context.wasPinned)",
-        "leafCount=\(context.leafCount)",
-        "newTreeEmpty=\(context.newTreeEmpty)",
-        "focusedSurfaceID=\(SupatermLog.uuid(context.focusedSurfaceID))",
-        "nextSurfaceID=\(SupatermLog.uuid(context.nextSurfaceID))",
-      ]
-    )
-  }
-
-  func logCloseKillSurface(
-    surfaceID: UUID,
-    tabID: TerminalTabID,
-    source: TerminalSurfaceCloseSource
-  ) {
-    SupatermLog.debug(
-      SupatermLog.terminal,
-      "terminal.close.killSurface",
-      fields: [
-        "source=\(source.rawValue)",
-        "surfaceID=\(SupatermLog.uuid(surfaceID))",
-        "tabID=\(SupatermLog.uuid(tabID.rawValue))",
-      ]
-    )
-  }
-
-  func logCloseRequest(
-    surfaceID: UUID,
-    needsConfirmationOverride: Bool?,
-    resolvedCloseRequest: ResolvedCloseRequest,
-    source: TerminalSurfaceCloseSource
-  ) {
-    let tabID = tabID(containing: surfaceID)
-    let resolvedTarget: String
-    let resolvedNeedsConfirmation: Bool
-    switch resolvedCloseRequest {
-    case .request(let request):
-      resolvedTarget = "\(request.target)"
-      resolvedNeedsConfirmation = request.needsConfirmation
-    case .window(let needsConfirmation):
-      resolvedTarget = "window"
-      resolvedNeedsConfirmation = needsConfirmation
-    }
-    SupatermLog.debug(
-      SupatermLog.terminal,
-      "terminal.close.request",
-      fields: [
-        "source=\(source.rawValue)",
-        "surfaceID=\(SupatermLog.uuid(surfaceID))",
-        "tabID=\(SupatermLog.uuid(tabID?.rawValue))",
-        "selectedTabID=\(SupatermLog.uuid(selectedTabID?.rawValue))",
-        "focusedSurfaceID=\(SupatermLog.uuid(tabID.flatMap { focusHistoryByTab[$0]?.current }))",
-        "isPinned=\(tabID.map { spaceManager.tab(for: $0)?.isPinned == true } ?? false)",
-        "needsConfirmationOverride=\(needsConfirmationOverride.map { "\($0)" } ?? "nil")",
-        "resolvedTarget=\(resolvedTarget)",
-        "resolvedNeedsConfirmation=\(resolvedNeedsConfirmation)",
-      ]
-    )
   }
 
   func emitFocusChangedIfNeeded(_ surfaceID: UUID) {
@@ -2496,34 +1314,6 @@ final class TerminalHostState {
     case .window(let needsConfirmation):
       emit(.windowCloseRequested(needsConfirmation: needsConfirmation))
     }
-  }
-
-  func mapGotoTabTarget(_ target: ghostty_action_goto_tab_e) -> TerminalGotoTabTarget? {
-    let raw = Int(target.rawValue)
-    if raw <= 0 {
-      switch raw {
-      case Int(GHOSTTY_GOTO_TAB_PREVIOUS.rawValue):
-        return .previous
-      case Int(GHOSTTY_GOTO_TAB_NEXT.rawValue):
-        return .next
-      case Int(GHOSTTY_GOTO_TAB_LAST.rawValue):
-        return .last
-      default:
-        return nil
-      }
-    }
-    return .index(raw)
-  }
-
-  func nextTabIndex(in spaceID: TerminalSpaceID) -> Int {
-    var maxIndex = 0
-    for tab in spaceManager.tabs(in: spaceID) {
-      guard tab.title.hasPrefix("Terminal ") else { continue }
-      let suffix = tab.title.dropFirst("Terminal ".count)
-      guard let value = Int(suffix) else { continue }
-      maxIndex = max(maxIndex, value)
-    }
-    return maxIndex + 1
   }
 
   func fallbackTitle(for tabID: TerminalTabID) -> String {
@@ -2617,338 +1407,6 @@ final class TerminalHostState {
       return surfaces[focusedSurfaceID]
     }
     return trees[tabID]?.root?.leftmostLeaf()
-  }
-
-  func observeSpaceCatalog() {
-    spaceCatalogObservationTask?.cancel()
-    spaceCatalogObservationTask = Task { @MainActor [weak self] in
-      let observations = Observations { [weak self] in
-        self?.spaceCatalog ?? .default
-      }
-      for await spaceCatalog in observations {
-        guard let self else { return }
-        self.applyObservedSpaceCatalog(spaceCatalog)
-      }
-    }
-  }
-
-  func applyObservedSpaceCatalog(_ spaceCatalog: TerminalSpaceCatalog) {
-    let resolvedSpaceCatalog = TerminalSpaceCatalog.sanitized(spaceCatalog)
-    guard resolvedSpaceCatalog != lastAppliedSpaceCatalog else { return }
-
-    let previousSelectedSpaceID = selectedSpaceID
-    lastAppliedSpaceCatalog = resolvedSpaceCatalog
-    let diff = spaceManager.applyCatalog(resolvedSpaceCatalog)
-    removeTrees(for: diff.removedTabIDs, source: .spaceCatalogObserved)
-    synchronizePinnedTabCatalogWithSpaces()
-
-    if previousSelectedSpaceID != selectedSpaceID {
-      finalizeSpaceSelectionChange()
-      sessionDidChange()
-    } else if !diff.removedTabIDs.isEmpty {
-      syncFocus(windowActivity)
-      sessionDidChange()
-    }
-  }
-
-  @discardableResult
-  func writeSpaceCatalog(
-    _ spaceCatalog: TerminalSpaceCatalog
-  ) -> TerminalSpaceManager.SpaceCatalogDiff {
-    let resolvedSpaceCatalog = TerminalSpaceCatalog.sanitized(spaceCatalog)
-    replaceSpaceCatalog(resolvedSpaceCatalog)
-    lastAppliedSpaceCatalog = resolvedSpaceCatalog
-
-    let diff = spaceManager.applyCatalog(resolvedSpaceCatalog)
-    removeTrees(for: diff.removedTabIDs, source: .spaceCatalogWrite)
-    synchronizePinnedTabCatalogWithSpaces()
-    return diff
-  }
-
-  func persistDefaultSelectedSpaceID(_ spaceID: TerminalSpaceID) {
-    guard spaceCatalog.defaultSelectedSpaceID != spaceID else { return }
-
-    var updatedSpaceCatalog = spaceCatalog
-    updatedSpaceCatalog.defaultSelectedSpaceID = spaceID
-    replaceSpaceCatalog(updatedSpaceCatalog)
-    lastAppliedSpaceCatalog = updatedSpaceCatalog
-  }
-
-  func replaceSpaceCatalog(_ spaceCatalog: TerminalSpaceCatalog) {
-    $spaceCatalog.withLock { $0 = spaceCatalog }
-  }
-
-  func nextSelectedSpaceID(afterDeleting spaceID: TerminalSpaceID) -> TerminalSpaceID {
-    let remainingSpaces = spaces.filter { $0.id != spaceID }
-    precondition(!remainingSpaces.isEmpty)
-
-    if let selectedSpaceID,
-      selectedSpaceID != spaceID,
-      remainingSpaces.contains(where: { $0.id == selectedSpaceID })
-    {
-      return selectedSpaceID
-    }
-
-    if let deletedIndex = spaces.firstIndex(where: { $0.id == spaceID }) {
-      for space in spaces[..<deletedIndex].reversed()
-      where remainingSpaces.contains(where: { $0.id == space.id }) {
-        return space.id
-      }
-    }
-
-    return remainingSpaces[0].id
-  }
-
-  func finalizeSpaceSelectionChange() {
-    guard managesTerminalSurfaces else {
-      lastEmittedFocusSurfaceID = nil
-      return
-    }
-    ensureInitialTab(focusing: false)
-    if let selectedTabID {
-      focusSurface(in: selectedTabID)
-    } else {
-      lastEmittedFocusSurfaceID = nil
-    }
-    syncFocus(windowActivity)
-  }
-
-  func removeTrees(
-    for tabIDs: [TerminalTabID],
-    terminateSessions: Bool = true,
-    source: TerminalTreeRemovalSource
-  ) {
-    for tabID in tabIDs {
-      removeTree(for: tabID, terminateSessions: terminateSessions, source: source)
-    }
-  }
-
-  func liveSurfaceIDs() -> [UUID] {
-    Array(surfaces.keys).sorted { $0.uuidString < $1.uuidString }
-  }
-
-  func killZmxSession(for surfaceID: UUID) {
-    killZmxSessions(for: [surfaceID])
-  }
-
-  func killZmxSessions(for surfaceIDs: [UUID]) {
-    let surfaceIDs = Array(Set(surfaceIDs))
-    guard !surfaceIDs.isEmpty else {
-      SupatermLog.debug(SupatermLog.zmx, "zmx.kill.skipped", fields: ["reason=empty"])
-      return
-    }
-    guard zmxSessionsEnabled else {
-      SupatermLog.debug(SupatermLog.zmx, "zmx.kill.skipped", fields: ["reason=disabled"])
-      return
-    }
-    guard zmxClient.isBundled() else {
-      SupatermLog.debug(SupatermLog.zmx, "zmx.kill.skipped", fields: ["reason=unbundled"])
-      return
-    }
-    SupatermLog.debug(
-      SupatermLog.zmx,
-      "zmx.kill.enqueue",
-      fields: [
-        "count=\(surfaceIDs.count)",
-        "surfaceIDs=\(Self.logSurfaceIDs(surfaceIDs))",
-      ]
-    )
-    let zmxClient = zmxClient
-    Task.detached(priority: .utility) {
-      await withTaskGroup(of: Void.self) { group in
-        for surfaceID in surfaceIDs {
-          group.addTask {
-            await zmxClient.killSession(surfaceID)
-          }
-        }
-      }
-    }
-  }
-
-  func killZmxSessionsAndWait(for surfaceIDs: [UUID]) async {
-    let surfaceIDs = Array(Set(surfaceIDs))
-    guard !surfaceIDs.isEmpty else {
-      SupatermLog.debug(SupatermLog.zmx, "zmx.killAndWait.skipped", fields: ["reason=empty"])
-      return
-    }
-    guard zmxSessionsEnabled else {
-      SupatermLog.debug(SupatermLog.zmx, "zmx.killAndWait.skipped", fields: ["reason=disabled"])
-      return
-    }
-    guard zmxClient.isBundled() else {
-      SupatermLog.debug(SupatermLog.zmx, "zmx.killAndWait.skipped", fields: ["reason=unbundled"])
-      return
-    }
-    SupatermLog.debug(
-      SupatermLog.zmx,
-      "zmx.killAndWait.start",
-      fields: [
-        "count=\(surfaceIDs.count)",
-        "surfaceIDs=\(Self.logSurfaceIDs(surfaceIDs))",
-      ]
-    )
-    let zmxClient = zmxClient
-    await withTaskGroup(of: Void.self) { group in
-      for surfaceID in surfaceIDs {
-        group.addTask {
-          await zmxClient.killSession(surfaceID)
-        }
-      }
-    }
-    SupatermLog.debug(
-      SupatermLog.zmx,
-      "zmx.killAndWait.finished",
-      fields: [
-        "count=\(surfaceIDs.count)",
-        "surfaceIDs=\(Self.logSurfaceIDs(surfaceIDs))",
-      ]
-    )
-  }
-
-  func terminateLiveTerminalSessions() {
-    killZmxSessions(for: liveSurfaceIDs())
-  }
-
-  func terminateLiveTerminalSessionsAndWait() async {
-    await killZmxSessionsAndWait(for: liveSurfaceIDs())
-  }
-
-  func mapSplitDirection(_ direction: GhosttySplitAction.NewDirection)
-    -> SplitTree<GhosttySurfaceView>.NewDirection
-  {
-    switch direction {
-    case .left:
-      return .left
-    case .right:
-      return .right
-    case .up:
-      return .up
-    case .down:
-      return .down
-    }
-  }
-
-  func mapSplitDirection(_ direction: TerminalPaneSplitDirection)
-    -> SplitTree<GhosttySurfaceView>.Direction
-  {
-    switch direction {
-    case .horizontal:
-      return .horizontal
-    case .vertical:
-      return .vertical
-    }
-  }
-
-  func mapSessionSplitDirection(
-    _ direction: SplitTree<GhosttySurfaceView>.Direction
-  ) -> TerminalPaneSplitDirection {
-    switch direction {
-    case .horizontal:
-      return .horizontal
-    case .vertical:
-      return .vertical
-    }
-  }
-
-  func mapPaneDirection(_ direction: SupatermPaneDirection)
-    -> SplitTree<GhosttySurfaceView>.NewDirection
-  {
-    switch direction {
-    case .down:
-      return .down
-    case .left:
-      return .left
-    case .right:
-      return .right
-    case .up:
-      return .up
-    }
-  }
-
-  func mapFocusDirection(_ direction: GhosttySplitAction.FocusDirection)
-    -> SplitTree<GhosttySurfaceView>.FocusDirection
-  {
-    switch direction {
-    case .previous:
-      return .previous
-    case .next:
-      return .next
-    case .left:
-      return .spatial(.left)
-    case .right:
-      return .spatial(.right)
-    case .up:
-      return .spatial(.up)
-    case .down:
-      return .spatial(.down)
-    }
-  }
-
-  func mapResizeDirection(_ direction: GhosttySplitAction.ResizeDirection)
-    -> SplitTree<GhosttySurfaceView>.SpatialDirection
-  {
-    switch direction {
-    case .left:
-      return .left
-    case .right:
-      return .right
-    case .up:
-      return .up
-    case .down:
-      return .down
-    }
-  }
-
-  func mapResizeDirection(_ direction: SupatermResizePaneDirection)
-    -> SplitTree<GhosttySurfaceView>.SpatialDirection
-  {
-    switch direction {
-    case .left:
-      return .left
-    case .right:
-      return .right
-    case .up:
-      return .up
-    case .down:
-      return .down
-    }
-  }
-
-  func mapPaneAxis(_ axis: SupatermPaneAxis)
-    -> SplitTree<GhosttySurfaceView>.Direction
-  {
-    switch axis {
-    case .horizontal:
-      return .horizontal
-    case .vertical:
-      return .vertical
-    }
-  }
-
-  func mapPaneSizeUnit(_ unit: SupatermPaneSizeUnit)
-    -> SplitTree<GhosttySurfaceView>.SizeUnit
-  {
-    switch unit {
-    case .cells:
-      return .cells
-    case .percent:
-      return .percent
-    }
-  }
-
-  func mapDropZone(_ zone: TerminalSplitTreeView.DropZone)
-    -> SplitTree<GhosttySurfaceView>.NewDirection
-  {
-    switch zone {
-    case .up:
-      return .up
-    case .bottom:
-      return .down
-    case .left:
-      return .left
-    case .right:
-      return .right
-    }
   }
 
   func debugPaneSnapshot(
